♻️ Changed button tokens structure
