+++ libraries/core-react/src/components/Button/Button.types.ts
@@ -1,0 +1,14 @@
+import type { ComponentToken } from '@equinor/eds-tokens'
+
+export type ButtonToken = ComponentToken & {
+  entities?: {
+    icon: ComponentToken
+  }
+}
+
+export type ButtonTokenSet = {
+  contained: ButtonToken
+  outlined: ButtonToken
+  ghost: ButtonToken
+  ghost_icon: ButtonToken
+}
